+++ RELEASES.md
@@ -1,4 +1,10 @@
-### Version 5.4.5
+## Version 5.5.0 - work in progress
+Tidy the creation of qualifiedName.
+
+Tidy up some of unnecessary methods in the model
+* Remove getParent(), relpace with parent attribute.
+
+## Version 5.4.5
 Fix issue where final state calls the error method instead of a warning.
 
 Refactor logging, wanrings and errors; remove need for casting in TypeScript and ensure the logging, wanring and error object passed will have the correct methods.